Move firestore utilities from utils/ to firebase/ directory
- Relocated firestore utility files to better match the project structure

- Updated all imports to reference the new location

- Verified the code compiles without errors
