+++ apps/web/src/hooks/firestore/useFirestoreCollection.ts
@@ -5,7 +5,7 @@ import {
   type FirestoreDocument,
   createEventCollectionFactory,
   getLatestEntityState,
-} from '../../utils/firestore';
+} from '../../firebase/firestore';
 
 /**
  * Reactフック: イベントソーシングパターンに基づくコレクションの取得と監視
